Refactor notification pipeline: centralize in Mediator
- Remove Notify methods from INotifiable; keep only DispatchNotifications
- Rewrite Notifier as a sealed, injectable singleton
- Move notification logic to Mediator; resolve handlers internally
- Refactor Mediator to handle all notification, command, and request logic
- Replace QuartzNotifier with QuartzMediator (IMediator implementation)
- Update Quartz job and tests to use IMediator
- Update NotifierMock and SpyNotifiable for new interface
- Use object initializer for Notifier/Mediator in tests
- Add BOM to project files
