+++ src/NetMediate.Core/INotifiable.cs
@@ -31,34 +31,4 @@ public interface INotifiable
         CancellationToken cancellationToken = default
     )
         where TMessage : notnull;
-
-    /// <summary>
-    /// Publishes a single notification through the pipeline and all registered handlers.
-    /// </summary>
-    /// <typeparam name="TMessage">The notification message type.</typeparam>
-    /// <param name="key">An optional key to distinguish this notification from others of the same message type.</param>
-    /// <param name="message">The notification message instance.</param>
-    /// <param name="cancellationToken">A token to observe for cancellation.</param>
-    /// <returns>A <see cref="Task"/> that completes when all handlers have been notified.</returns>
-    Task Notify<TMessage>(
-        object? key,
-        TMessage message,
-        CancellationToken cancellationToken = default
-    )
-        where TMessage : notnull;
-
-    /// <summary>
-    /// Publishes each message in the sequence through the pipeline and all registered handlers.
-    /// </summary>
-    /// <typeparam name="TMessage">The notification message type.</typeparam>
-    /// <param name="key">An optional key to distinguish this notification from others of the same message type.</param>
-    /// <param name="messages">The sequence of notification messages to publish.</param>
-    /// <param name="cancellationToken">A token to observe for cancellation.</param>
-    /// <returns>A <see cref="Task"/> that completes when all messages have been dispatched.</returns>
-    Task Notify<TMessage>(
-        object? key,
-        IEnumerable<TMessage> messages,
-        CancellationToken cancellationToken = default
-    )
-        where TMessage : notnull;
 }
