+++ src/NetMediate.Quartz/QuartzNotifier.cs
@@ -1,7 +1,5 @@
-using System.Diagnostics.CodeAnalysis;
 using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
-using Quartz;
+using System.Diagnostics.CodeAnalysis;
 
 namespace NetMediate.Quartz;
 
@@ -12,8 +10,8 @@ namespace NetMediate.Quartz;
 [RequiresUnreferencedCode(
     "QuartzNotificationJob uses reflection to resolve message types by name and dispatch notifications."
 )]
-[DecoratorFor<INotifiable>(Order = int.MinValue)]
-internal sealed class QuartzNotifier : INotifiable
+[DecoratorFor<INotifiable>]
+public sealed class QuartzNotifier : INotifiable
 {
     /// <summary>
     /// Gets or sets the underlying <see cref="INotifiable"/> instance that this decorator wraps.
@@ -21,79 +19,11 @@ internal sealed class QuartzNotifier : INotifiable
     /// <remarks>This property is required to fulfill the contract of <see cref="INotifiable"/> and enables
     /// the use of this class as a decorator. The property may not be used directly in typical scenarios.</remarks>
     [Inject] public required INotifiable Inner { get; init; }
-    [Inject] internal IScheduler Scheduler { get; init; }
-    [Inject] internal INotificationSerializer Serializer { get; init; }
-    [Inject] internal IOptions<QuartzNotificationOptions> Options { get; init; }
+
+    /// <summary>
+    /// Gets the logger for QuartzNotifier operations.
+    /// </summary>
     [Inject] internal ILogger<QuartzNotifier> Logger { get; init; }
-
-    /// <inheritdoc />
-    public async Task Notify<TMessage>(
-        object? key,
-        TMessage message,
-        CancellationToken cancellationToken = default
-    )
-        where TMessage : notnull
-    {
-        var json = Serializer.Serialize(message);
-        var typeName =
-            typeof(TMessage).AssemblyQualifiedName
-            ?? throw new InvalidOperationException(
-                $"Cannot determine assembly-qualified name for type '{typeof(TMessage).FullName}'."
-            );
-
-        var jobKey = new JobKey($"{typeof(TMessage).Name}_{Guid.NewGuid():N}", Options.Value.GroupName);
-
-        var jobBuilder = JobBuilder
-            .Create<QuartzNotificationJob>()
-            .WithIdentity(jobKey)
-            .UsingJobData(QuartzNotificationJob.MessageDataKey, json)
-            .UsingJobData(QuartzNotificationJob.TypeDataKey, typeName)
-            .StoreDurably(false);
-
-        if (key is not null)
-        {
-            jobBuilder = jobBuilder
-                .UsingJobData(
-                    QuartzNotificationJob.KeyDataKey,
-                    System.Text.Json.JsonSerializer.Serialize(key)
-                )
-                .UsingJobData(
-                    QuartzNotificationJob.KeyTypeDataKey,
-                    key.GetType().AssemblyQualifiedName ?? key.GetType().FullName ?? "System.Object"
-                );
-        }
-
-        var job = jobBuilder.Build();
-
-        var trigger = TriggerBuilder
-            .Create()
-            .WithIdentity($"{jobKey.Name}_trigger", Options.Value.GroupName)
-            .StartNow()
-            .Build();
-
-        await Scheduler.ScheduleJob(job, trigger, cancellationToken).ConfigureAwait(false);
-
-        if (Logger.IsEnabled(LogLevel.Debug))
-        {
-            Logger.LogDebug(
-                "QuartzNotifier: scheduled notification job {JobKey} for message type {MessageType}.",
-                jobKey,
-                typeof(TMessage).Name
-            );
-        }
-    }
-
-    /// <inheritdoc />
-    public async Task Notify<TMessage>(
-        object? key,
-        IEnumerable<TMessage> messages,
-        CancellationToken cancellationToken = default
-    )
-        where TMessage : notnull
-    {
-        foreach (var message in messages)
-            await Notify(key, message, cancellationToken).ConfigureAwait(false);
-    }
 
     /// <inheritdoc />
     public async Task DispatchNotifications<TMessage>(
@@ -112,7 +42,6 @@ internal sealed class QuartzNotifier : INotifiable
             );
         }
 
-        foreach (var handler in handlers)
-            await handler.Handle(message, cancellationToken).ConfigureAwait(false);
+        await Inner.DispatchNotifications(key, message, handlers, cancellationToken).ConfigureAwait(false);
     }
 }
