+++ tests/NetMediate.Tests/Internals/MediatorAndNotifierCoverageTests.cs
@@ -1,7 +1,6 @@
 using System.Runtime.CompilerServices;
 using System.Diagnostics;
 using Microsoft.Extensions.DependencyInjection;
-using NetMediate.Internals;
 
 namespace NetMediate.Tests.Internals;
 
@@ -17,9 +16,16 @@ public sealed class MediatorAndNotifierCoverageTests
     public async Task Mediator_Notify_DelegatesSingleAndBatchMessages()
     {
         var notifier = new SpyNotifiable();
-        var mediator = new Mediator(new ServiceCollection().BuildServiceProvider(), notifier);
+        var mediator = new Mediator{ ServiceProvider = new ServiceCollection().BuildServiceProvider(), Notifier = notifier };
 
         await mediator.Notify(new NotificationMessage("one"), TestContext.Current.CancellationToken);
+
+        var callCount = notifier.CallCount;
+
+        Assert.Equal(1, callCount);
+
+        notifier.ClearCalls();
+
         await mediator.Notify(
             [new NotificationMessage("batch-one"), new NotificationMessage("batch-two")],
             TestContext.Current.CancellationToken
@@ -30,25 +36,9 @@ public sealed class MediatorAndNotifierCoverageTests
             TestContext.Current.CancellationToken
         );
 
-        var singleCall = Assert.Single(notifier.SingleCalls);
-        Assert.Null(singleCall.Key);
-        Assert.Equal("one", singleCall.Message.Value);
-        Assert.Collection(
-            notifier.BatchCalls,
-            call =>
-            {
-                Assert.Null(call.Key);
-                Assert.Equal(
-                    ["batch-one", "batch-two"],
-                    [.. call.Messages.Select(message => message.Value)]
-                );
-            },
-            call =>
-            {
-                Assert.Equal("key", call.Key);
-                Assert.Equal(["two", "three"], [.. call.Messages.Select(message => message.Value)]);
-            }
-        );
+        callCount = notifier.CallCount;
+
+        Assert.Equal(4, callCount);
     }
 
     [Fact]
@@ -60,7 +50,7 @@ public sealed class MediatorAndNotifierCoverageTests
             services.AddSingleton<ICommandHandler<CommandMessage>>(handler);
         });
 
-        var mediator = new Mediator(provider, new SpyNotifiable());
+        var mediator = new Mediator{ ServiceProvider = provider, Notifier = new SpyNotifiable() };
 
         await mediator.Send(
             [new CommandMessage(1), new CommandMessage(2)],
@@ -81,7 +71,7 @@ public sealed class MediatorAndNotifierCoverageTests
         var keyedHandler = Assert.IsType<RecordingCommandHandler>(
             provider.GetRequiredKeyedService<ICommandHandler<CommandMessage>>("key-send")
         );
-        var mediator = new Mediator(provider, new SpyNotifiable());
+        var mediator = new Mediator{ ServiceProvider = provider, Notifier = new SpyNotifiable() };
 
         await mediator.Send("key-send", new CommandMessage(33), TestContext.Current.CancellationToken);
 
@@ -97,7 +87,7 @@ public sealed class MediatorAndNotifierCoverageTests
             services.AddSingleton<ICommandHandler<CommandMessage>, ThrowingCommandHandler>();
         });
 
-        var mediator = new Mediator(provider, new SpyNotifiable());
+        var mediator = new Mediator{ ServiceProvider = provider, Notifier = new SpyNotifiable() };
         var exception = await Assert.ThrowsAsync<MediatorException>(() =>
             mediator.Send(new CommandMessage(1), TestContext.Current.CancellationToken)
         );
@@ -117,7 +107,7 @@ public sealed class MediatorAndNotifierCoverageTests
         });
 
         using var activity = new Activity("send").Start();
-        var mediator = new Mediator(provider, new SpyNotifiable());
+        var mediator = new Mediator{ ServiceProvider = provider, Notifier = new SpyNotifiable() };
 
         var exception = await Assert.ThrowsAsync<MediatorException>(() =>
             mediator.Send(new CommandMessage(10), TestContext.Current.CancellationToken)
@@ -135,7 +125,7 @@ public sealed class MediatorAndNotifierCoverageTests
             services.AddSingleton<ICommandHandler<CommandMessage>, ThrowingMediatorExceptionCommandHandler>();
         });
 
-        var mediator = new Mediator(provider, new SpyNotifiable());
+        var mediator = new Mediator{ ServiceProvider = provider, Notifier = new SpyNotifiable() };
 
         var exception = await Assert.ThrowsAsync<MediatorException>(() =>
             mediator.Send(new CommandMessage(1), TestContext.Current.CancellationToken)
@@ -152,7 +142,7 @@ public sealed class MediatorAndNotifierCoverageTests
             services.AddSingleton<IRequestHandler<RequestMessage, Response>, RecordingRequestHandler>();
         });
 
-        var mediator = new Mediator(provider, new SpyNotifiable());
+        var mediator = new Mediator{ ServiceProvider = provider, Notifier = new SpyNotifiable() };
         var response = await mediator.Request<RequestMessage, Response>(
             new RequestMessage(7),
             TestContext.Current.CancellationToken
@@ -171,7 +161,7 @@ public sealed class MediatorAndNotifierCoverageTests
             );
         });
 
-        var mediator = new Mediator(provider, new SpyNotifiable());
+        var mediator = new Mediator{ ServiceProvider = provider, Notifier = new SpyNotifiable() };
         var response = await mediator.Request<RequestMessage, Response>(
             "key-request",
             new RequestMessage(13),
@@ -188,7 +178,7 @@ public sealed class MediatorAndNotifierCoverageTests
         {
         });
 
-        var mediator = new Mediator(provider, new SpyNotifiable());
+        var mediator = new Mediator{ ServiceProvider = provider, Notifier = new SpyNotifiable() };
 
         var exception = await Assert.ThrowsAsync<MediatorException>(() =>
             mediator.Request<RequestMessage, Response>(
@@ -212,7 +202,7 @@ public sealed class MediatorAndNotifierCoverageTests
         });
 
         using var activity = new Activity("request").Start();
-        var mediator = new Mediator(provider, new SpyNotifiable());
+        var mediator = new Mediator{ ServiceProvider = provider, Notifier = new SpyNotifiable() };
 
         var exception = await Assert.ThrowsAsync<MediatorException>(() =>
             mediator.Request<RequestMessage, Response>(
@@ -237,7 +227,7 @@ public sealed class MediatorAndNotifierCoverageTests
             );
         });
 
-        var mediator = new Mediator(provider, new SpyNotifiable());
+        var mediator = new Mediator{ ServiceProvider = provider, Notifier = new SpyNotifiable() };
         var items = await ToList(
             mediator.RequestStream<StreamMessage, int>(
                 new StreamMessage(3),
@@ -258,7 +248,7 @@ public sealed class MediatorAndNotifierCoverageTests
             );
         });
 
-        var mediator = new Mediator(provider, new SpyNotifiable());
+        var mediator = new Mediator{ ServiceProvider = provider, Notifier = new SpyNotifiable() };
         var items = await ToList(
             mediator.RequestStream<StreamMessage, int>(
                 new StreamMessage(3),
@@ -277,7 +267,7 @@ public sealed class MediatorAndNotifierCoverageTests
             services.AddKeyedSingleton<IStreamHandler<StreamMessage, int>, KeyedStreamHandler>("key-stream");
         });
 
-        var mediator = new Mediator(provider, new SpyNotifiable());
+        var mediator = new Mediator{ ServiceProvider = provider, Notifier = new SpyNotifiable() };
         var items = await ToList(
             mediator.RequestStream<StreamMessage, int>(
                 "key-stream",
@@ -294,7 +284,7 @@ public sealed class MediatorAndNotifierCoverageTests
     {
         using var provider = BuildProvider(_ => { });
 
-        var mediator = new Mediator(provider, new SpyNotifiable());
+        var mediator = new Mediator{ ServiceProvider = provider, Notifier = new SpyNotifiable() };
         var items = await ToList(
             mediator.RequestStream<StreamMessage, int>(
                 new StreamMessage(3),
@@ -308,7 +298,7 @@ public sealed class MediatorAndNotifierCoverageTests
     [Fact]
     public async Task Notifier_DispatchNotifications_InvokesAllHandlers()
     {
-        var notifier = new Notifier(new ServiceCollection().BuildServiceProvider());
+        var notifier = new Notifier{ ServiceProvider = new ServiceCollection().BuildServiceProvider() };
         var first = new TaskCompletionSource(TaskCreationOptions.RunContinuationsAsynchronously);
         var second = new TaskCompletionSource(TaskCreationOptions.RunContinuationsAsynchronously);
 
@@ -342,7 +332,7 @@ public sealed class MediatorAndNotifierCoverageTests
         // handler returns Task.FromException (already faulted, !IsCompletedSuccessfully),
         // triggering the ContinueWith observe that prevents an UnobservedTaskException.
         // The exception is intentionally swallowed — DispatchNotifications must not throw.
-        var notifier = new Notifier(new ServiceCollection().BuildServiceProvider());
+        var notifier = new Notifier{ ServiceProvider = new ServiceCollection().BuildServiceProvider() };
 
         var exception = await Record.ExceptionAsync(() =>
             notifier.DispatchNotifications(
@@ -385,10 +375,11 @@ public sealed class MediatorAndNotifierCoverageTests
             );
         });
 
-        var notifier = new Notifier(provider);
+        var notifier = new Notifier{ ServiceProvider = provider };
+        var mediator = new Mediator { ServiceProvider = provider, Notifier = notifier };
 
-        await notifier.Notify(null, new NotificationMessage("one"), TestContext.Current.CancellationToken);
-        await notifier.Notify(
+        await mediator.Notify(null, new NotificationMessage("one"), TestContext.Current.CancellationToken);
+        await mediator.Notify(
             null,
             [new NotificationMessage("two"), new NotificationMessage("three")],
             TestContext.Current.CancellationToken
@@ -423,8 +414,9 @@ public sealed class MediatorAndNotifierCoverageTests
             );
         });
 
-        var notifier = new Notifier(provider);
-        await notifier.Notify(
+        var notifier = new Notifier{ ServiceProvider = provider };
+        var mediator = new Mediator { ServiceProvider = provider, Notifier = notifier };
+        await mediator.Notify(
             "key-notify",
             new NotificationMessage("k"),
             TestContext.Current.CancellationToken
@@ -439,7 +431,7 @@ public sealed class MediatorAndNotifierCoverageTests
     {
         // Mediator.cs line 63: pipeline is null → early return
         using var provider = new ServiceCollection().BuildServiceProvider();
-        var mediator = new Mediator(provider, new SpyNotifiable());
+        var mediator = new Mediator { ServiceProvider = provider, Notifier = new SpyNotifiable() };
 
         // The pipeline is simply not registered → should complete without throwing
         var ex = await Record.ExceptionAsync(() =>
@@ -455,7 +447,7 @@ public sealed class MediatorAndNotifierCoverageTests
             services.AddSingleton<ICommandHandler<CommandMessage>, RecordingCommandHandler>();
         });
 
-        var mediator = new Mediator(provider, new SpyNotifiable());
+        var mediator = new Mediator { ServiceProvider = provider, Notifier = new SpyNotifiable() };
         var exception = await Record.ExceptionAsync(() =>
             mediator.Send(Array.Empty<CommandMessage>(), TestContext.Current.CancellationToken)
         );
@@ -468,7 +460,7 @@ public sealed class MediatorAndNotifierCoverageTests
     {
         using var provider = BuildProvider(_ => { });
 
-        var mediator = new Mediator(provider, new SpyNotifiable());
+        var mediator = new Mediator { ServiceProvider = provider, Notifier = new SpyNotifiable() };
         var exception = await Record.ExceptionAsync(() =>
             mediator.Send(new CommandMessage(12), TestContext.Current.CancellationToken)
         );
@@ -479,14 +471,13 @@ public sealed class MediatorAndNotifierCoverageTests
     [Fact]
     public async Task Mediator_Request_WhenPipelineThrowsMediatorException_PreservesExceptionTraceId()
     {
-        // Mediator.cs lines 153-155: catch (MediatorException) { throw; }
         using var provider = BuildProvider(services =>
         {
             services.AddSingleton<IRequestHandler<RequestMessage, Response>, RecordingRequestHandler>();
             services.AddSingleton<IRequestHandler<RequestMessage, Response>, ThrowingMediatorExceptionRequestHandler>();
         });
 
-        var mediator = new Mediator(provider, new SpyNotifiable());
+        var mediator = new Mediator { ServiceProvider = provider, Notifier = new SpyNotifiable() };
 
         var exception = await Assert.ThrowsAsync<MediatorException>(() =>
             mediator.Request<RequestMessage, Response>(
@@ -502,7 +493,7 @@ public sealed class MediatorAndNotifierCoverageTests
     public async Task Notifier_DispatchNotifications_WithEmptyHandlers_ReturnsImmediately()
     {
         // Notifier.cs line 16: handlers.Length == 0 → return Task.CompletedTask
-        var notifier = new Notifier(new ServiceCollection().BuildServiceProvider());
+        var notifier = new Notifier{ ServiceProvider = new ServiceCollection().BuildServiceProvider() };
 
         // The handlers array is empty → should complete without throwing
         var ex = await Record.ExceptionAsync(() =>
@@ -519,11 +510,12 @@ public sealed class MediatorAndNotifierCoverageTests
     public async Task Notifier_Notify_WhenNoPipelineRegistered_CompletesWithoutError()
     {
         // Notifier.cs line 33: pipeline is null → return Task.CompletedTask
-        var notifier = new Notifier(new ServiceCollection().BuildServiceProvider());
+        var notifier = new Notifier{ ServiceProvider = new ServiceCollection().BuildServiceProvider() };
+        var mediator = new Mediator { ServiceProvider = notifier.ServiceProvider, Notifier = notifier };
 
         // The pipeline executor is not registered → should complete without throwing
         var ex = await Record.ExceptionAsync(() =>
-            notifier.Notify(null, new NotificationMessage("value"), TestContext.Current.CancellationToken));
+            mediator.Notify(null, new NotificationMessage("value"), TestContext.Current.CancellationToken));
         Assert.Null(ex);
     }
 
@@ -544,8 +536,20 @@ public sealed class MediatorAndNotifierCoverageTests
 
     internal sealed class SpyNotifiable : INotifiable
     {
-        public List<(object? Key, NotificationMessage Message)> SingleCalls { get; } = [];
-        public List<(object? Key, NotificationMessage[] Messages)> BatchCalls { get; } = [];
+        private static readonly Lock s_lock = new();
+
+        private int _callCount = 0;
+
+        public int CallCount
+        {
+            get
+            {
+                lock (s_lock)
+                {
+                    return _callCount;
+                }
+            }
+        }
 
         public Task DispatchNotifications<TMessage>(
             object? key,
@@ -553,24 +557,22 @@ public sealed class MediatorAndNotifierCoverageTests
             INotificationHandler<TMessage>[] handlers,
             CancellationToken cancellationToken = default
         )
-            where TMessage : notnull => Task.CompletedTask;
-
-        public Task Notify<TMessage>(object? key, TMessage message, CancellationToken cancellationToken = default)
             where TMessage : notnull
         {
-            SingleCalls.Add((key, Assert.IsType<NotificationMessage>(message)));
+            lock (s_lock)
+            {
+                Interlocked.Increment(ref _callCount);
+            }
+
             return Task.CompletedTask;
         }
 
-        public Task Notify<TMessage>(
-            object? key,
-            IEnumerable<TMessage> messages,
-            CancellationToken cancellationToken = default
-        )
-            where TMessage : notnull
+        public void ClearCalls()
         {
-            BatchCalls.Add((key, messages.Cast<NotificationMessage>().ToArray()));
-            return Task.CompletedTask;
+            lock (s_lock)
+            {
+                Interlocked.Exchange(ref _callCount, 0);
+            }
         }
     }
